moves firebase modules to server
since we only the server is using this module
