+++ server/firebase-private-key.js
