+++ server/package.json
@@ -6,6 +6,7 @@
   "author": "cs472-spots",
   "license": "MIT",
   "dependencies": {
+    "dotenv": "^4.0.0",
     "ejs": "^2.5.5",
     "express": "^4.14.1",
     "firebase-admin": "4.1.1",
